requirements.txt
requests
beautifulsoup4
jinja2
python-dateutil
lxml
